Updated benchmark

+++ src/benchmark.cpp
@@ -10,8 +10,8 @@
 #include <chrono>
 using namespace std;
 
-// Taken from https://www.gormanalysis.com/blog/reading-and-writing-csv-files-with-cpp/
-void write_csv(std::string filename, std::vector<std::pair<std::string, std::vector<int>>> dataset){
+// Modified from https://www.gormanalysis.com/blog/reading-and-writing-csv-files-with-cpp/
+void write_csv(std::string filename, std::vector<std::pair<std::string, std::vector<double>>> dataset){
     // Make a CSV file with one or more columns of integer values
     // Each column of data is represented by the pair <column name, column data>
     //   as std::pair<std::string, std::vector<int>>
@@ -59,12 +59,12 @@ int main(int argc, char **argv) {
     srand (1);
     const int64_t num_ops = 500;
     const int64_t megabyte = 1 << 20;
-    vector<int32_t> inputDataSize = {1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024};
+    vector<double> inputDataSize = {1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 1024};
 
     cerr << "Running benchmarking..." << endl;
-    vector<int32_t> put_tps_Binary;
-    vector<int32_t> get_tps_Binary;
-    vector<int32_t> scan_tps_Binary;
+    vector<double> put_tps_Binary;
+    vector<double> get_tps_Binary;
+    vector<double> scan_tps_Binary;
 
     cerr << "Creating Database with size: " << constants::MEMTABLE_SIZE << " memtable entries...";
     Database db(constants::MEMTABLE_SIZE);
@@ -115,7 +115,7 @@ int main(int argc, char **argv) {
     
 
     // Wrap into a vector
-    std::vector<std::pair<std::string, std::vector<int>>> vals = {{"InputDataSize", inputDataSize},  
+    std::vector<std::pair<std::string, std::vector<double>>> vals = {{"InputDataSize", inputDataSize},  
                                                                   {"Put_Binary", put_tps_Binary}, {"Get_Binary", get_tps_Binary}, {"Scan_Binary", scan_tps_Binary}};
     
     // Write the vector to CSV
